Send and receive config message
The motivation for this patch is due to the change in the KNoT AMQP API,
which changed the name of config to event and unified the event and
schema messages in a new type of message called config.
The new KNoT AMQP API can be found at
https://github.com/CESARBR/knot-babeltower on the file /docs/events.md,
using the hash 4aa5a91.

This patch makes the following changes:
- Change the API for sending and receiving a config message instead of
a schema;
- Adapts the content of the config message to match the new AMQP API of
KNoT;
- Change the parser to work with the new json messages of the new
KNoT AMQP API.